requirements.txt
numpy==2.4.3
opencv-python==4.13.0.92
openpyxl==3.1.5
pandas==3.0.1
python-dotenv==1.2.2
tqdm==4.67.3
